worker/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
requests
google-cloud-storage
pydantic==2.5.0
pydantic-settings==2.1.0
tenacity==8.2.3

# AI Core
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cu121
huggingface_hub>=0.23.2
diffusers==0.27.2
transformers==4.44.0
accelerate==0.34.0
controlnet-aux==0.0.9
safetensors>=0.4.3
opencv-python-headless==4.9.0.80
insightface==0.7.3
onnxruntime-gpu==1.16.3
peft==0.7.1

